+++ context/lb_gnb.go
@@ -7,12 +7,12 @@ import (
 )
 
 type LbGnb struct{
-	GnbID 		int
+	GnbID 		int64
 	LbConn 		*LBConn
 	Ues 		sync.Map
 }
 
-func NewLbGnb(id int) (gnb *LbGnb){
+func NewLbGnb(id int64) (gnb *LbGnb){
 	gnb.GnbID = id
 	gnb.LbConn = NewLBConn()
 	gnb.LbConn.ID = id
@@ -20,11 +20,11 @@ func NewLbGnb(id int) (gnb *LbGnb){
 	return gnb
 }
 
-func (gnb *LbGnb) AddAMFUe(id int) {
+func (gnb *LbGnb) AddAMFUe(id int64) {
 	gnb.Ues.Store(id, NewUE(id))
 }
 
-func (gnb *LbGnb) ContainsUE(id int) (cont bool) {
+func (gnb *LbGnb) ContainsUE(id int64) (cont bool) {
 	_, cont = gnb.Ues.Load(id)
 	return 
 }
